refactor: consolidate modal and side panel types into CrmTypes.ts and remove unused files
