+++ frontend/src/community/crm/v2/enums/common.ts
@@ -5,9 +5,9 @@ export enum CrmPriorityEnum {
 }
 
 export enum CrmTaskTabEnum {
-  MY_TASKS = "my-tasks",
-  ALL_TASKS = "all-tasks",
-  COMPLETED_TASKS = "completed-tasks"
+  MY_TASKS = "my tasks",
+  ALL_TASKS = "all tasks",
+  COMPLETED_TASKS = "completed tasks"
 }
 
 export enum CrmIndustryEnum {
