--- frontend/src/community/crm/v2/types/ModalTypes.ts
@@ -1,14 +1,0 @@
-export enum CrmModalTypes {
-  ADD_COMPANY_MODAL = "ADD_COMPANY_MODAL",
-  ADD_CONTACT_MODAL = "ADD_CONTACT_MODAL",
-  EDIT_COMPANY_MODAL = "EDIT_COMPANY_MODAL",
-  EDIT_CONTACT_MODAL = "EDIT_CONTACT_MODAL",
-  ADD_TASK_MODAL = "ADD_TASK_MODAL",
-  EDIT_TASK_MODAL = "EDIT_TASK_MODAL",
-  DELETE_COMPANY_MODAL = "DELETE_COMPANY_MODAL",
-  DELETE_CONTACT_MODAL = "DELETE_CONTACT_MODAL",
-  ADD_DEAL_STAGE_MODAL = "ADD_DEAL_STAGE_MODAL",
-  EDIT_DEAL_STAGE_MODAL = "EDIT_DEAL_STAGE_MODAL",
-  DELETE_DEAL_STAGE_MODAL = "DELETE_DEAL_STAGE_MODAL",
-  DELETE_TASK_MODAL = "DELETE_TASK_MODAL"
-}
