--- frontend/src/community/crm/v2/types/SidePanelTypes.ts
@@ -1,7 +1,0 @@
-export enum CrmSidePanelTypes {
-  CONTACT_SIDE_PANEL = "CONTACT_SIDE_PANEL",
-  COMPANY_SIDE_PANEL = "COMPANY_SIDE_PANEL",
-  TASK_SIDE_PANEL = "TASK_SIDE_PANEL",
-  ADD_DEAL_SIDE_PANEL = "ADD_DEAL_SIDE_PANEL",
-  DEAL_DETAIL_SIDE_PANEL = "DEAL_DETAIL_SIDE_PANEL"
-}
